An initial implementation of the abstract tuple / abstract field feature
There are still some missing cases or undesirable semantics or unclear error
messages but they will be considered in the future. For now the feature is
good enough. Added a couple of tricky examples.

Some things to consider in the future: (1) the evaluation order is potentially
unintuitive; (2) private fields need to be supported.