Simplify evaluation environment
It turns out that the evaluation does not need the state monad, just the reader
monad. This is because incredibly, the semantics of this language does not
require any persistent state to be kept. All environment needed for evaluation
can just exist as a function argument and be discarded after the function call
finishes.