Make the evaluation order more intuitive
Previously, if an abstract tuple's abstract field is overridden mutiple times,
the user gets an uninformative error message because the evaluation order of
that field happens after other fields.

Now, such a field retains the original evaluation order.

Arguably, it may be even more intuitive if it is evaluated multiple times with
different values. Alas, a case could also be made that this behavior is less
intuitive.

Consider this example:

```
abstract { abstract a, ret = a / 2 } { a = 42 } { a = 64 }.eval
```

Should the result be equivalent to `{a = 64, ret = 32}`?
Or should be it `{a = 64, ret = 21}`?