Make the handling of private and final rows more uniform
Instead of having the `TupleValueRowPrivacy` and `TupleValueRowFinality` be
separate types and fields, just keep the entire `RowAttribute` in the evaluated
value.

Also improves error messages to make the private field error as
informative as the final field override error.

Add a few more examples illustrating the use of private fields.